backend/requirements.txt
Django
djangorestframework
django-cors-headers
google-generativeai
python-dotenv